+++ src/lib/form_email.js
@@ -30,7 +30,7 @@ module.exports.checkFormEmail = (value, type, options) => {
   if (options.required === true && (result.value === undefined || result.value === ``)) {
     const err = {
       code: 101,
-      error: `The value for the ${type} was empty.`,
+      error: `The value for "${type}" was empty.`,
     };
     result.errors.push(err);
     result.errstr += `${err.error}\r\n`;
@@ -47,7 +47,7 @@ module.exports.checkFormEmail = (value, type, options) => {
   } catch (ex) {
     const err = {
       code: 103,
-      error: `An exception error occurred while attempting to parse the ${type} as an array.`,
+      error: `An exception error occurred while attempting to parse "${type}" as an array.`,
     };
     result.errors.push(err);
     result.errstr += `${err.error}\r\n`;
@@ -62,7 +62,7 @@ module.exports.checkFormEmail = (value, type, options) => {
           if (result.valueAsArray[i].length > Number(options.maxlength)) {
             const err = {
               code: 104,
-              error: `The value for the ${type} - '${result.valueAsArray[i]}' - failed to meet the specification for the maxlengthimum number of characters (${options.maxlength}).`,
+              error: `The value for "${type}" - '${result.valueAsArray[i]}' - failed to meet the specification for the maximum number of characters (${options.maxlength}).`,
             };
             result.errors.push(err);
             result.errstr += `${err.error}\r\n`;
@@ -73,7 +73,7 @@ module.exports.checkFormEmail = (value, type, options) => {
         if (result.value.length > Number(options.maxlength)) {
           const err = {
             code: 104,
-            error: `The value for the ${type} failed to meet the specification for the maxlengthimum number of characters (${options.maxlength}).`,
+            error: `The value for "${type}" failed to meet the specification for the maximum number of characters (${options.maxlength}).`,
           };
           result.errors.push(err);
           result.errstr += `${err.error}\r\n`;
@@ -83,7 +83,7 @@ module.exports.checkFormEmail = (value, type, options) => {
   } catch (ex) {
     const err = {
       code: 105,
-      error: `An exception error occurred while attempting to check if the ${type} met the specification for the maxlengthimum number of characters.`,
+      error: `An exception error occurred while attempting to check if "${type}" met the specification for the maximum number of characters.`,
       exception: ex.message,
     };
     result.errors.push(err);
@@ -99,7 +99,7 @@ module.exports.checkFormEmail = (value, type, options) => {
           if (result.valueAsArray[i].length < Number(options.minlength)) {
             const err = {
               code: 106,
-              error: `The value for the ${type} - '${result.valueAsArray[i]}' - failed to meet the specification for the minlengthimum number of characters (${options.minlength}).`,
+              error: `The value for "${type}" - '${result.valueAsArray[i]}' - failed to meet the specification for the minimum number of characters (${options.minlength}).`,
             };
             result.errors.push(err);
             result.errstr += `${err.error}\r\n`;
@@ -110,7 +110,7 @@ module.exports.checkFormEmail = (value, type, options) => {
         if (result.value.length < Number(options.minlength)) {
           const err = {
             code: 106,
-            error: `The value for the ${type} failed to meet the specification for the minlengthimum number of characters (${options.minlength}).`,
+            error: `The value for "${type}" failed to meet the specification for the minimum number of characters (${options.minlength}).`,
           };
           result.errors.push(err);
           result.errstr += `${err.error}\r\n`;
@@ -120,7 +120,7 @@ module.exports.checkFormEmail = (value, type, options) => {
   } catch (ex) {
     const err = {
       code: 107,
-      error: `An exception error occurred while attempting to check if the ${type} met the specification for the minlengthimum number of characters.`,
+      error: `An exception error occurred while attempting to check if "${type}" met the specification for the minimum number of characters.`,
       exception: ex.message,
     };
     result.errors.push(err);
@@ -136,7 +136,7 @@ module.exports.checkFormEmail = (value, type, options) => {
           if (result.valueAsArray[i].length === Number(options.size)) {
             const err = {
               code: 108,
-              error: `The value for the ${type} - '${result.valueAsArray[i]}' - failed to meet the specification for the required number of characters (${options.size}).`,
+              error: `The value for "${type}" - '${result.valueAsArray[i]}' - failed to meet the specification for the required number of characters (${options.size}).`,
             };
             result.errors.push(err);
             result.errstr += `${err.error}\r\n`;
@@ -147,7 +147,7 @@ module.exports.checkFormEmail = (value, type, options) => {
         if (result.value.length === Number(options.size)) {
           const err = {
             code: 108,
-            error: `The value for the ${type} failed to meet the specification for the required number of characters (${options.size}).`,
+            error: `The value for "${type}" failed to meet the specification for the required number of characters (${options.size}).`,
           };
           result.errors.push(err);
           result.errstr += `${err.error}\r\n`;
@@ -157,7 +157,7 @@ module.exports.checkFormEmail = (value, type, options) => {
   } catch (ex) {
     const err = {
       code: 109,
-      error: `An exception error occurred while attempting to check if the ${type} met the specification for the required number of characters.`,
+      error: `An exception error occurred while attempting to check if "${type}" met the specification for the required number of characters.`,
       exception: ex.message,
     };
     result.errors.push(err);
@@ -174,7 +174,7 @@ module.exports.checkFormEmail = (value, type, options) => {
           if (options.pattern.test(result.valueAsArray[i]) === false) {
             const err = {
               code: 110,
-              error: `The value for the ${type} - '${result.valueAsArray[i]}' - failed to meet the specified pattern requirements.`,
+              error: `The value for "${type}" - '${result.valueAsArray[i]}' - failed to meet the specified pattern requirements.`,
             };
             result.errors.push(err);
             result.errstr += `${err.error}\r\n`;
@@ -185,7 +185,7 @@ module.exports.checkFormEmail = (value, type, options) => {
         if (options.pattern.test(result.value) === false) {
           const err = {
             code: 110,
-            error: `The value for the ${type} failed to meet the specified pattern requirements.`,
+            error: `The value for "${type}" failed to meet the specified pattern requirements.`,
           };
           result.errors.push(err);
           result.errstr += `${err.error}\r\n`;
@@ -195,7 +195,7 @@ module.exports.checkFormEmail = (value, type, options) => {
   } catch (ex) {
     const err = {
       code: 111,
-      error: `An exception error occurred while attempting to check if the ${type} met the specified pattern requirements.`,
+      error: `An exception error occurred while attempting to check if "${type}" met the specified pattern requirements.`,
       exception: ex.message,
     };
     result.errors.push(err);